+++ modules/ingest-common/src/yamlRestTest/resources/rest-api-spec/test/ingest/300_redact_processor.yml
@@ -1,0 +1,39 @@
+---
+"Test put redact pipeline":
+  - do:
+      ingest.put_pipeline:
+        id: "pipeline-using-a-redact-processor"
+        body:  >
+          {
+            "processors": [
+              {
+                "redact": {
+                  "field": "to_redact",
+                  "patterns": ["%{EMAILADDRESS:EMAIL}", "%{IP:IP_ADDRESS}", "%{CREDIT_CARD:CREDIT_CARD}"],
+                  "pattern_definitions": {
+                    "CREDIT_CARD": "\\d{4}[ -]\\d{4}[ -]\\d{4}[ -]\\d{4}"
+                  }
+                }
+              }
+            ]
+          }
+
+---
+"Test ignore missing":
+  - do:
+      ingest.simulate:
+        body:  >
+          {
+            "pipeline": {
+            "processors": [
+              {
+                "redact": {
+                  "field": "to_redact",
+                  "ignore_missing": false,
+                  "patterns": ["%{EMAILADDRESS:EMAIL}", "%{IP:IP_ADDRESS}"]
+                }
+              }
+            ]},
+            "docs": [{"_source": {"wrong_field": "will error"}}]
+          }
+  - match: { docs.0.error.reason: "field [to_redact] is null or missing" }
