+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/MachineLearning.java
@@ -327,7 +327,6 @@ import org.elasticsearch.xpack.ml.inference.persistence.TrainedModelProvider;
 import org.elasticsearch.xpack.ml.inference.pytorch.process.BlackHolePyTorchProcess;
 import org.elasticsearch.xpack.ml.inference.pytorch.process.NativePyTorchProcessFactory;
 import org.elasticsearch.xpack.ml.inference.pytorch.process.PyTorchProcessFactory;
-import org.elasticsearch.xpack.ml.ingest.RedactProcessor;
 import org.elasticsearch.xpack.ml.job.JobManager;
 import org.elasticsearch.xpack.ml.job.JobManagerHolder;
 import org.elasticsearch.xpack.ml.job.NodeLoadDetector;
@@ -551,11 +550,8 @@ public class MachineLearning extends Plugin
             parameters.ingestService.getClusterService(),
             this.settings
         );
-        parameters.ingestService.addIngestClusterStateListener(inferenceFactory);
-        // TODO use a non-noop watchdog
-        RedactProcessor.Factory redactFactory = new RedactProcessor.Factory(parameters.client, MatcherWatchdog.noop());
 
-        return Map.of(InferenceProcessor.TYPE, inferenceFactory, RedactProcessor.TYPE, redactFactory);
+        return Map.of(InferenceProcessor.TYPE, inferenceFactory);
     }
 
     // This is not used in v8 and higher, but users are still prevented from setting it directly to avoid confusion
